src/data_train/requirements.txt
pandas>=2.2.1
matplotlib>=3.8.4
lightgbm>=4.3.0
scikit-learn>=1.4.2
pyyaml>=6.0.1
numpy>=1.26.4
python-dotenv>=1.0.1
mlflow-skinny>=2.15.1